Add examples of a number of large javascript widgets: mapping, rich text editor, syntax highlighting.
Because I'm hoping to create a plugin framework put each of these into their own ASDF project.

Plot-window-examples is loads them all and defines a little demo function.

Change things to conform to the typical javascript approach of having a single global in the window,
in our case this window.bah, which consists of an object with presumably useful things in it.  So
far that includes: the websocket stream, the message sending function, and a logging function.

Log4cl logging is more widely used, and i'm now letting it name the loggers more.

A very casually written log4cl appender can be found in the epilog of main.lisp, really
just a doddle.  I'm playing with the idea of logging to an in memory data structure where
I can then browse analyse etc. after the fact.