Introduce an interpolate macro in both lisp and parenscript.  In the Lisp case it invokes the code inside of cl-interpol; this avoids the use of readmacro.  I don't like readmacros.  In Parnscript it calls the cl-interpol code to expand the interpolated string into Lisp code; it then walks that code to rewrite it into Parenscript.  Needly to say this works only for a very limited subset of cl-interpol's functionality; but it's still quite useful.  Also add optima to our ASDF.
Introduce a parenscript macro cprogn that can be when a series of continuations are needed to do something.  This was a fun idea which I probably didn't actuall need, time will tell.

Introduce with-new-element, which is a precursor of an insertion point for adding wdgets to the page.  Getting the animation on this thing to work was, ah, educational.

Misc other junk that probably ought to be stripped back out.